+++ Sprint1/TecSus - API/API/.gitignore
@@ -1,0 +1,1 @@
+/target/
