+++ Sprint1/TecSus - API/API/pom.xml
@@ -28,6 +28,11 @@
 			<version>3.0.0</version>
 		</dependency>
 		<dependency>
+			<groupId>io.springfox</groupId>
+			<artifactId>springfox-swagger2</artifactId>
+			<version>3.0.0</version>
+		</dependency>
+		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-web</artifactId>
 		</dependency>
@@ -44,7 +49,7 @@
 		<dependency>
 		  <groupId>org.postgresql</groupId>
 		  <artifactId>postgresql</artifactId>
-		  <version>42.2.2</version>
+		  <version>42.2.2</version><!--$NO-MVN-MAN-VER$-->
 		</dependency>
 		
 		<dependency>
