+++ Sprint1/TecSus - API/API/src/main/java/com/tecsus/API/config/SwaggerConfig.java
@@ -1,0 +1,47 @@
+package com.tecsus.API.config;
+
+import java.util.ArrayList;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.service.ApiInfo;
+import springfox.documentation.service.Contact;
+import springfox.documentation.service.VendorExtension;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+import static springfox.documentation.builders.PathSelectors.regex;
+@Configuration
+@EnableSwagger2
+public class SwaggerConfig {
+
+	@Bean
+	public Docket productApi() {
+		return new Docket (DocumentationType.SWAGGER_2)
+				.select()
+				.apis(RequestHandlerSelectors.basePackage("com.tecsus.API"))
+				.paths(PathSelectors.any())
+				.build()
+				.apiInfo(metaInfo());
+	}
+	
+	private ApiInfo metaInfo() {
+		ApiInfo apiInfo = new ApiInfo(
+				"TecSus API REST",
+				"API REST",
+				"1.0",
+				"Terms of Service",
+				new Contact("PowerTech","https://github.com/luks-ecdc/PowerTech-API-TecSUS","null"),
+				
+				"Licença",//há colocar
+				"null",new ArrayList<VendorExtension>()
+				);
+				
+				return apiInfo;
+	}
+	
+
+}
