+++ Sprint1/TecSus - API/API/src/main/java/com/tecsus/API/controller/UnidadeController.java
@@ -17,8 +17,8 @@ import com.tecsus.API.entities.Unidade;
 import com.tecsus.API.repository.UnidadeRepository;
 
 @RestController // Permiti spring reconhecer como uma controller	
-@CrossOrigin // impede conflito de CORS, assim o frontend consegue receber o back
-@RequestMapping// defini url que vai consultar a entidade
+@CrossOrigin(origins="*") // impede conflito de CORS, assim o frontend consegue receber o back
+@RequestMapping(value = "/unidade")// defini url que vai consultar a entidade
 public class UnidadeController {
 	
 	@Autowired
