+++ Sprint1/TecSus - API/API/src/main/java/com/tecsus/API/entities/Unidade.java
@@ -1,0 +1,95 @@
+package com.tecsus.API.entities;
+
+import java.io.Serializable;
+import java.util.Objects;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+
+@Entity // Mapea a classe como uma tabela no banco de dados
+@Table(name = "tb_unidade")
+public class Unidade implements Serializable{
+
+	private static final long serialVersionUID = 1L;
+	
+	@Id
+	@GeneratedValue(strategy = GenerationType.AUTO) // Incrementa o ID automaticamente
+	private Long id;
+	private String nome;
+	private int cpf;
+	private int cep;
+	private String endereco;
+	private String complemento ;
+	
+	public Unidade() {}
+	
+	public Unidade(Long id, String nome, int cpf, int cep, String endereco, String complemento) {
+		super();
+		this.id = id;
+		this.nome = nome;
+		this.cpf = cpf;
+		this.cep = cep;
+		this.endereco = endereco;
+		this.complemento = complemento;
+	}
+	// Inicio Construtores 
+	
+	@Override
+	public int hashCode() {
+		return Objects.hash(id);
+	}
+	@Override
+	public boolean equals(Object obj) {
+		if (this == obj)
+			return true;
+		if (obj == null)
+			return false;
+		if (getClass() != obj.getClass())
+			return false;
+		Unidade other = (Unidade) obj;
+		return Objects.equals(id, other.id);
+	}
+	public Long getId() {
+		return id;
+	}
+	public void setId(Long id) {
+		this.id = id;
+	}
+	public String getNome() {
+		return nome;
+	}
+	public void setNome(String nome) {
+		this.nome = nome;
+	}
+	public int getCpf() {
+		return cpf;
+	}
+	public void setCpf(int cpf) {
+		this.cpf = cpf;
+	}
+	public int getCep() {
+		return cep;
+	}
+	public void setCep(int cep) {
+		this.cep = cep;
+	}
+	public String getEndereco() {
+		return endereco;
+	}
+	public void setEndereco(String endereco) {
+		this.endereco = endereco;
+	}
+	public String getComplemento() {
+		return complemento;
+	}
+	public void setComplemento(String complemento) {
+		this.complemento = complemento;
+	}
+	
+	
+			
+}
