+++ Sprint1/TecSus - API/API/src/main/java/com/tecsus/API/entities/Usuario.java
@@ -1,0 +1,100 @@
+package com.tecsus.API.entities;
+
+import java.io.Serializable;
+import java.util.Objects;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+import com.tecsus.API.entities.enums.Funcao;
+
+@Entity // Mapea a classe como uma tabela no banco de dados
+@Table(name = "tb_usuario")
+public class Usuario implements Serializable{
+
+	private static final long serialVersionUID = 1L;
+	
+	@Id
+	@GeneratedValue(strategy = GenerationType.AUTO) // Incrementa o ID automaticamente
+	private Long id;
+	private String username;
+	private Funcao funcao;
+	private String email;
+	private String senha;
+	private int perfil ;
+	
+	
+	// Inicio Construtores 
+	
+	public Usuario() {}
+	 
+	public Long getId() {
+		return id;
+	}
+	public void setId(Long id) {
+		this.id = id;
+	}
+	public String getUsername() {
+		return username;
+	}
+	public void setUsername(String username) {
+		this.username = username;
+	}
+	public Funcao getFuncao() {
+		return funcao;
+	}
+	public void setFuncao(Funcao funcao) {
+		this.funcao = funcao;
+	}
+	public String getEmail() {
+		return email;
+	}
+	public void setEmail(String email) {
+		this.email = email;
+	}
+	public String getSenha() {
+		return senha;
+	}
+	public void setSenha(String senha) {
+		this.senha = senha;
+	}
+	public int getPerfil() {
+		return perfil;
+	}
+	public void setPerfil(int perfil) {
+		this.perfil = perfil;
+	}
+
+	public Usuario(Long id, String username, Funcao funcao, String email, String senha, int perfil) {
+		super();
+		this.id = id;
+		this.username = username;
+		this.funcao = funcao;
+		this.email = email;
+		this.senha = senha;
+		this.perfil = perfil;
+	}
+	
+	// Fim Construtores 
+	
+	@Override
+	public int hashCode() {
+		return Objects.hash(id);
+	}
+
+	@Override
+	public boolean equals(Object obj) {
+		if (this == obj)
+			return true;
+		if (obj == null)
+			return false;
+		if (getClass() != obj.getClass())
+			return false;
+		Usuario other = (Usuario) obj;
+		return Objects.equals(id, other.id);
+	}
+		
+}
