+++ Sprint1/TecSus - API/API/src/main/resources/application.properties
@@ -1,9 +1,9 @@
 spring.jpa.database=POSTGRESQL
 spring.sql.init.platform=postgres
-spring.datasource.url= jdbc:postgresql://localhost:5432/postgres
+spring.datasource.url= jdbc:postgresql://localhost:5434/postgres
 spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
 spring.datasource.username=postgres
-spring.datasource.password=123456
+spring.datasource.password=123
 spring.jpa.show-sql=true
 spring.jpa.generate-ddl=true
 spring.jpa.hibernate.ddl-auto=update
